feat: allow different rep schemes to be used
also split weekProgrammingHelpers into smaller files
